Fix encryption modals for mobile
